Move Kubernetes cluster service to its own module
