+++ extensions/kubernetes-cluster-service/runtime/pom.xml
@@ -21,30 +21,36 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-kubernetes-parent</artifactId>
+        <artifactId>camel-quarkus-kubernetes-cluster-service-parent</artifactId>
         <version>3.10.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>camel-quarkus-kubernetes-deployment</artifactId>
-    <name>Camel Quarkus :: Kubernetes :: Deployment</name>
+    <artifactId>camel-quarkus-kubernetes-cluster-service</artifactId>
+    <name>Camel Quarkus :: Kubernetes Cluster Service :: Runtime</name>
+    <description>Provides a Kubernetes implementation of the Camel Cluster Service SPI</description>
+
+    <properties>
+        <camel.quarkus.jvmSince>3.10.0</camel.quarkus.jvmSince>
+        <camel.quarkus.nativeSince>3.10.0</camel.quarkus.nativeSince>
+    </properties>
 
     <dependencies>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-kubernetes-client-deployment</artifactId>
+            <artifactId>quarkus-kubernetes-client</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-core-deployment</artifactId>
+            <artifactId>camel-quarkus-core</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-core-cloud-deployment</artifactId>
+            <artifactId>camel-quarkus-core-cloud</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-support-commons-logging-deployment</artifactId>
+            <artifactId>camel-quarkus-support-commons-logging</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
@@ -54,6 +60,10 @@
 
     <build>
         <plugins>
+            <plugin>
+                <groupId>io.quarkus</groupId>
+                <artifactId>quarkus-extension-maven-plugin</artifactId>
+            </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
@@ -70,4 +80,31 @@
         </plugins>
     </build>
 
+    <profiles>
+        <profile>
+            <id>full</id>
+            <activation>
+                <property>
+                    <name>!quickly</name>
+                </property>
+            </activation>
+            <build>
+                <plugins>
+                    <plugin>
+                        <groupId>org.apache.camel.quarkus</groupId>
+                        <artifactId>camel-quarkus-maven-plugin</artifactId>
+                        <executions>
+                            <execution>
+                                <id>update-extension-doc-page</id>
+                                <goals>
+                                    <goal>update-extension-doc-page</goal>
+                                </goals>
+                                <phase>process-classes</phase>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 </project>
