+++ extensions/kubernetes-cluster-service/runtime/src/main/doc/configuration.adoc
@@ -1,0 +1,20 @@
+ifeval::[{doc-show-extra-content} == true]
+include::extra-content.adoc[]
+endif::[]
+
+=== Having only a single consumer in a cluster consuming from a given endpoint
+
+When the same route is deployed on multiple pods, it could be interesting to use this extension in conjunction with the xref:reference/extensions/master.adoc[Master one].
+In such a setup, a single consumer will be active at a time across the whole camel master namespace.
+
+For instance, having the route below deployed on multiple pods:
+
+```
+from("master:ns:timer:test?period=100").log("Timer invoked on a single pod at a time");
+```
+
+As a result, a single consumer will be active across the `ns` camel master namespace.
+It means that, at a given time, only a single timer will generate exchanges across the whole cluster.
+In other words, messages will be logged every 100ms on a single pod at a time.
+
+The kubernetes cluster service could further be tuned by tweaking `quarkus.camel.cluster.kubernetes.*` properties.
