+++ extensions/kubernetes-cluster-service/runtime/src/main/java/org/apache/camel/quarkus/component/kubernetes/cluster/KubernetesClusterServiceBuildTimeConfig.java
@@ -24,12 +24,15 @@ import io.quarkus.runtime.annotations.ConfigRoot;
 
 @ConfigRoot(name = "camel.cluster.kubernetes", phase = ConfigPhase.BUILD_AND_RUN_TIME_FIXED)
 public class KubernetesClusterServiceBuildTimeConfig {
-
     /**
      * Whether a Kubernetes Cluster Service should be automatically configured
      * according to 'quarkus.camel.cluster.kubernetes.*' configurations.
+     *
+     * @deprecated this property is no longer needed as the Kubernetes implementation of the Camel CLuster Service API has
+     *             been moved to a dedicated extension.
      */
-    @ConfigItem(defaultValue = "false")
+    @Deprecated(since = "3.10.0", forRemoval = true)
+    @ConfigItem(defaultValue = "true")
     public boolean enabled;
 
     /**
@@ -47,5 +50,4 @@ public class KubernetesClusterServiceBuildTimeConfig {
             return config.enabled;
         }
     }
-
 }
