+++ integration-tests/master-openshift/src/main/resources/application.properties
@@ -17,5 +17,4 @@
 
 application.id = leader
 
-quarkus.camel.cluster.kubernetes.enabled = true
 quarkus.camel.cluster.kubernetes.rebalancing = false
